// legacy/schema.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
      admins: {
        Row: {
          created_at: string
          id: string
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          user_id: string
        }
        Update: {
          created_at?: string
          id?: string
          user_id?: string
        }
        Relationships: []
      }
      brands: {
        Row: {
          created_at: string
          id: string
          logo_url: string | null
          name: string
          slug: string
        }
        Insert: {
          created_at?: string
          id?: string
          logo_url?: string | null
          name: string
          slug: string
        }
        Update: {
          created_at?: string
          id?: string
          logo_url?: string | null
          name?: string
          slug?: string
        }
        Relationships: []
      }
      categories: {
        Row: {
          created_at: string
          id: string
          image_url: string
          slug: string
          title: string
        }
        Insert: {
          created_at?: string
          id?: string
          image_url: string
          slug: string
          title: string
        }
        Update: {
          created_at?: string
          id?: string
          image_url?: string
          slug?: string
          title?: string
        }
        Relationships: []
      }
      colour: {
        Row: {
          colour_name: string
          created_at: string
          id: string
        }
        Insert: {
          colour_name: string
          created_at?: string
          id?: string
        }
        Update: {
          colour_name?: string
          created_at?: string
          id?: string
        }
        Relationships: []
      }
      frame_styles: {
        Row: {
          created_at: string
          id: string
          slug: string
          title: string
        }
        Insert: {
          created_at?: string
          id?: string
          slug: string
          title: string
        }
        Update: {
          created_at?: string
          id?: string
          slug?: string
          title?: string
        }
        Relationships: []
      }
// ...
      products: {
        Row: {
          attributes: { [key: string]: string }[] | null
          brand: string
          category: string
          created_at: string
          description: string
          featured: boolean | null
          frame_style: string | null
          gender: string | null
          has_variants: boolean
          id: string
          images: string[]
          instock: boolean | null
          price: number
          slug: string
          sub_category: string
          title: string
          variants: { variant_name: string; variant_value: string }[] | null
        }
        Insert: {
          attributes?: { [key: string]: string }[] | null
          brand: string
          category: string
          created_at?: string
          description: string
          featured?: boolean | null
          frame_style?: string | null
          gender?: string | null
          has_variants?: boolean
          id?: string
          images: string[] | null
          instock?: boolean | null
          price?: number
          slug: string
          sub_category: string
          title: string
          variants?: Json | null
        }
        Update: {
          attributes?: { [key: string]: string }[] | null
          brand?: string
          category?: string
          created_at?: string
          description?: string
          featured?: boolean | null
          frame_style?: string | null
          gender?: string | null
          has_variants?: boolean
          id?: string
          images?: string[] | null
          instock?: boolean | null
          price?: number
          slug?: string
          sub_category?: string
          title?: string
          variants?: Json | null
        }
        Relationships: [
          {
            foreignKeyName: "products_brand_fkey"
            columns: ["brand"]
            isOneToOne: false
            referencedRelation: "brands"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "products_category_fkey"
            columns: ["category"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "products_sub_category_fkey"
            columns: ["sub_category"]
            isOneToOne: false
            referencedRelation: "sub_categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      size: {
        Row: {
          created_at: string
          id: string
          size_name: string
        }
        Insert: {
          created_at?: string
          id?: string
          size_name: string
        }
        Update: {
          created_at?: string
          id?: string
          size_name?: string
        }
        Relationships: []
      }
      sub_categories: {
        Row: {
          category: string
          created_at: string
          id: string
          image_url: string
          slug: string
          title: string
        }
        Insert: {
          category: string
          created_at?: string
          id?: string
          image_url: string
          slug: string
          title: string
        }
        Update: {
          category?: string
          created_at?: string
          id?: string
          image_url?: string
          slug?: string
          title?: string
        }
        Relationships: [
          {
            foreignKeyName: "sub_categories_category_fkey"
            columns: ["category"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
// ...
}